backup/requirements.txt
tweepy==4.15.0
praw==7.7.1
beautifulsoup4==4.12.2
Pillow==10.1.0
requests==2.31.0
argparse==1.4.0